fix: ensure `next/font`(s) are applied correctly
